+++ README.md
@@ -77,6 +77,16 @@ Instead, it will copy all the configuration files and the transitive dependencie
 
 You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
 
+## Development
+
+### ENV
+
+You may create a `.env.local` file to overload some environment variables defined in other `.env` files. See [production env](.env.production) for example.
+
+### Doc
+
+For libraries and components, see [the doc for dev](dev-doc.md)
+
 ## Learn More
 
 You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
