+++ dev-doc.md
@@ -1,0 +1,7 @@
+# Development Guide
+
+## lib
+
+### `requests.js`
+
+Send HTTP requests with this library.
