+++ src/lib/index.js
@@ -1,0 +1,1 @@
+export { get, post, put } from './request';
